Analytisc with HyperQueue implementation
A resource-agnostic (HPC or local) implementation for monolingual data
analytics. In theory, this implementation should run seamlessly, the
existence of HPC or not should only change the way HyperQueue is set up.
In the case of local just starting a hq server and a hq worker. In the
case of HPC, start a hq server and a hq allocation queue.

All the procedures that were part of the map have been grouped into a
single script. Parallelization should be line-based, preferably, not
procedure-based, which makes it more complicated.

The reduce has also been grouped into a single script, but each one on
its own bash function, so all can be wrapped into the same pipeline:
read concat each batch of the map, apply function, write to tmp, then
move to permanent.

readcorpus and readdocument positional arguments have been reodered, so
the call can be just script.py <lang> and will use stdin/stdout by
default.

The singularity container to run each task has been added and the
Dockerfile for it has some GPU stuff commented out for now (or not).